the identifier refactoring, adjustments in generating names for introduced vars in graph controller
